added edit action to Tasks
